backend/requirements.txt
Django==2.1.2
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
MarkupSafe==1.1.1
oauthlib==3.1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
requirementslib==1.5.3
virtualenv==20.0.20
whitenoise==3.3.1
dj-database-url==0.4.2
psycopg2-binary==2.8.5